+++ ios/RNBackgroundGeolocation/TSLocationManager.framework/Headers/TSLocationManager.h
@@ -54,8 +54,8 @@
 - (void) removeGeofence:(NSString*)identifier success:(void (^)(NSString*))success error:(void (^)(NSString*))error;
 - (void) removeGeofences:(NSArray*)identifiers success:(void (^)(NSString*))success error:(void (^)(NSString*))error;;
 - (NSArray*) getGeofences;
-- (void) updateCurrentPosition:(NSDictionary*)options;
-- (void) watchPosition:(NSDictionary*)options;
+- (void) getCurrentPosition:(NSDictionary*)options success:(void (^)(NSDictionary*))success failure:(void (^)(NSError*))failure;
+- (void) watchPosition:(NSDictionary*)options success:(void (^)(NSDictionary*))success failure:(void (^)(NSError*))failure;
 - (void) stopWatchPosition;
 - (void) playSound:(SystemSoundID)soundId;
 - (void) notify:(NSString*)message;
@@ -68,7 +68,7 @@
 - (void) emailLog:(NSString*)to;
 - (void) setLogLevel:(NSInteger)level;
 - (CLLocationDistance)getOdometer;
-- (void) resetOdometer;
+- (void) setOdometer:(CLLocationDistance)odometer success:(void (^)(NSDictionary*))success failure:(void (^)(NSError*))failure;
 
 @end